+++ src/Faker/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Runtime.CompilerServices;
+
+// Provide access for testing purposes:
+[assembly: InternalsVisibleTo("RimuTec.Faker.Tests")]